Add tests for keep_related_entity_ids
